onnx/requirements.txt
onnx>=1.12.0
onnxruntime==1.22.0
opencv-python>=4.5.4.60
torch==1.11.0
torchvision==0.12.0
numpy>=1.23.5
pillow>=9.0.0
tqdm>=4.64.0
swanlab==0.5.7
